Ajuste de excepciones

+++ arnuv/src/main/java/com/core/arnuv/model/Seguridadpolitica.java
@@ -5,7 +5,6 @@ import lombok.Data;
 import org.hibernate.annotations.Comment;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -24,31 +23,31 @@ public class Seguridadpolitica implements Serializable {
 
 	@Comment("Longitud minima del password")
 	@Column(name = "longitud", precision = 2)
-	private BigDecimal longitud;
+	private Integer longitud;
 
 	@Comment("Numero de ingresos de password erroneo para bloquear al usuario")
 	@Column(name = "intentos", precision = 2)
-	private BigDecimal intentos;
+	private Integer intentos;
 
 	@Comment("Canridad minima de numeros que requiere el password")
 	@Column(name = "numeros", precision = 2)
-	private BigDecimal numeros;
+	private Integer numeros;
 
 	@Comment("Numero de caracteres minimos que requeiere el password")
 	@Column(name = "especiales", precision = 2)
-	private BigDecimal especiales;
+	private Integer especiales;
 
 	@Comment("Cantidad de letras en minusculas que requeire el password")
 	@Column(name = "minusculas", precision = 2)
-	private BigDecimal minusculas;
+	private Integer minusculas;
 
 	@Comment("Cantidad de letras en mayusculas que requiere el password")
 	@Column(name = "mayusculas", precision = 2)
-	private BigDecimal mayusculas;
+	private Integer mayusculas;
 
 	@Comment("Tiempo de Regeracion delToken en minutos")
 	@Column(name = "tiemporegeraciontoken", precision = 3)
-	private BigDecimal tiemporegeraciontoken;
+	private Integer tiemporegeraciontoken;
 
 	@OneToMany(mappedBy = "idpolitica")
 	private List<Rol> rols;
